admission: move tokensLinearModel into its own file
Pure code movement.

Release note: None
